refactor planet test analyssi
