+++ ignite/pkg/cosmosanalysis/module/testdata/planet_v2/proto/planet/planet.proto
@@ -1,11 +1,11 @@
 syntax = "proto3";
 package tendermint.planet.planet;
 import "google/api/annotations.proto";
-option go_package = "github.com/tendermint/planet/x/planet/types";
+option go_package = "github.com/tendermint/mars/x/mars/types";
 
 service Query {
   rpc MyQuery(QueryMyQueryRequest) returns (QueryMyQueryResponse) {
-    option (google.api.http).get = "/tendermint/planet/withoutmsg/my_query/{mytypefield}";
+    option (google.api.http).get = "/tendermint/mars/withoutmsg/my_query/{mytypefield}";
   }
 }
